+++ app/src/main/java/ru/pyrovsergey/gallerya3/model/pojo/User.java
@@ -4,10 +4,10 @@ import java.util.Objects;
 
 public class User {
 
-    private long id;
+    private int id;
     private String name;
 
-    public User(long id, String name) {
+    public User(int id, String name) {
         this.id = id;
         this.name = name;
     }
